+++ Gppbox/app/app.cpp
@@ -48,7 +48,7 @@ int main()
     //sf::RenderWindow window(sf::VideoMode(800, 600,32), "SFML works!");
     //sf::RenderWindow window(sf::VideoMode(1280, 720,32), "SFML works!");
 	window.setVerticalSyncEnabled(true);
-	window.setFramerateLimit(C::F_TARGET);
+	window.setFramerateLimit((unsigned int)C::F_TARGET);
 	Font font;
 
     if (!font.loadFromFile("res/MAIAN.TTF")) {
